Support cells with CellMetadata and ValueMetadata
In the OpenXML spec, there are three additional XML attributes that
may be present: 'cm', 'vm', and 'ph'. Before this commit, this would
raise an error specifying "Unknown attribute". After this commit, this
would continue to process the cell contents successfully without
raising.

The Cell Metadata index isn't required for rendering the data within the
cell. Neither is Value Metadata. Both of these are for extensions to
operate upon the cells.

The cell "Show Phonetic" attribute isn't required either, as it only
informs the tooltip the user would see if this is present.

These three additional attributes are ignored by Xlsxir, and won't fail
parsing upon finding them.